made message record count an int instead of a short
git-svn-id: http://svn.code.sf.net/p/dnsjava/code/trunk@350 c76caeb1-94fd-44dd-870f-0c9d92034fc1

+++ org/xbill/DNS/Cache.java
@@ -300,7 +300,7 @@ addMessage(Message in) {
 	short queryClass = in.getQuestion().getDClass();
 	byte cred;
 	short rcode = in.getHeader().getRcode();
-	short ancount = in.getHeader().getCount(Section.ANSWER);
+	int ancount = in.getHeader().getCount(Section.ANSWER);
 
 	if (rcode != Rcode.NOERROR && rcode != Rcode.NXDOMAIN)
 		return;
